+++ tests/ZestTest.php
@@ -180,6 +180,14 @@ class ZestTest extends \PHPUnit\Framework\TestCase {
 			// CSS escapes
 			[ "#cite_note-13\\.3F_It_Can\\'t_Be\\!-3", [ '/html[1]/body[1]/footer[1]/a[2]' ] ],
 			[ '#\\a9', [ '/html[1]/body[1]/footer[1]/small[1]/a[1]' ] ],
+			// CSS escapes that look like pseudoclasses (no match, but the selector is valid)
+			[ 'x\\:x', [] ],
+			[ '#x\\:x', [] ],
+			[ '.x\\:x', [] ],
+			// Escaped selectors of this form will match where appropriate
+			[ 'mw\\:section', [ "/html[1]/body[1]/mw:section[1]" ] ],
+			[ '#mw\\:section', [ "/html[1]/body[1]/mw:section[1]" ] ],
+			[ '.mw\\:class', [ "/html[1]/body[1]/mw:section[1]" ] ],
 			// The comma combinator
 			[ '#\\00a9, article p', [ '/html[1]/body[1]/article[1]/p[1]', '/html[1]/body[1]/footer[1]/small[1]/a[1]' ] ],
 			[ 'article > header, header + p', [ "/html[1]/body[1]/article[1]/header[1]", "/html[1]/body[1]/article[1]/p[1]" ] ],